pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
)/
'''

[tool.isort]
profile = "black"

[tool.mypy]
strict = true
disable_error_code = ["no-untyped-call", "no-any-return"]
python_version = "3.11"
plugins = [
  "sqlalchemy.ext.mypy.plugin",
  "pydantic.mypy"
]

[[tool.mypy.overrides]]
module = [
    "OSMPythonTools.nominatim",
    "apscheduler.*"
]
ignore_missing_imports = true
warn_return_any = false

[tool.poetry]
name = "open-prices"
version = "0.1.0"
description = "An open database of product prices"
authors = ["Open Food Facts <contact@openfoodfacts.org>"]
license = "AGPL-3.0 licence"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
Babel = "~2.13.1"
openfoodfacts = { version = "~2.4.0", extras = ["ml", "redis"] }
psycopg2-binary = "~2.9.9"
python-multipart = "~0.0.7"
requests = "~2.31.0"
osmpythontools = "^0.3.5"
typer = "~0.9.0"
django = "^5.0.7"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
django-filter = "^24.2"
factory-boy = "^3.3.0"
django-q2 = "^1.6.2"
croniter = "^3.0.3"
blessed = "^1.20.0"
ipython = "^8.26.0"
gunicorn = "^22.0.0"
django-cors-headers = "^4.4.0"
sentry-sdk = {extras = ["django"], version = "^2.13.0"}
django-solo = "^2.3.0"
pillow = "^10.4.0"
google-generativeai = "^0.8.3"

[tool.poetry.group.dev.dependencies]
black = "~23.12.1"
flake8 = "^7.0.0"
httpx = "^0.25.1"
isort = "^5.12.0"
mdx-truly-sane-lists = "^1.3"
mkdocs = "~1.5.3"
mkdocs-awesome-pages-plugin = "~2.9.2"
mkdocs-material = "~9.4.10"
mypy = "^1.8.0"
pre-commit = "^3.5.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
toml-sort = "^0.23.1"
types-requests = "^2.31.0.20240125"
types-tqdm = "^4.66.0.20240106"
types-sqlalchemy-utils = "^1.0.1"
pyparsing = "^3.1.4"
pydot = "^3.0.1"
freezegun = "^1.5.1"
